:package: [CHORE] 소셜 로그인 관련 폴더 구조 변경 #13
