+++ src/main/java/org/sopt/lequuServer/domain/user/dto/request/SocialLoginRequestDto.java
@@ -3,7 +3,7 @@ package org.sopt.lequuServer.domain.user.dto.request;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.sopt.lequuServer.global.sociallogin.SocialPlatform;
+import org.sopt.lequuServer.domain.user.model.SocialPlatform;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
