+++ src/main/java/org/sopt/lequuServer/global/auth/fegin/kakao/KakaoAuthApiClient.java
@@ -1,6 +1,6 @@
-package org.sopt.lequuServer.global.sociallogin.kakao.fegin;
+package org.sopt.lequuServer.global.auth.fegin.kakao;
 
-import org.sopt.lequuServer.global.sociallogin.kakao.response.KakaoAccessTokenResponse;
+import org.sopt.lequuServer.global.auth.fegin.kakao.response.KakaoAccessTokenResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PostMapping;
