+++ src/main/java/org/sopt/lequuServer/global/auth/fegin/kakao/response/KakaoUserResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.lequuServer.global.sociallogin.kakao.response;
+package org.sopt.lequuServer.global.auth.fegin.kakao.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
